refactor(developer): move message tests into test-messages.ts
